requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
numpy==1.19.2
opencv-python==4.4.0.42
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
pyttsx3==2.90
pytz==2020.1
pywin32==228
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
